refactor(container): make padded_string use pointer-width sizes
I want to be more consistent with what size type our containers use. Switch
padded_string to use ptrdiff_t like bump_vector and span use.

On x86_64, this improves performance slightly:

    Benchmark                                     Time             CPU      Time Old      Time New       CPU Old       CPU New
    --------------------------------------------------------------------------------------------------------------------------
    benchmark_parse_file                       -0.0098         -0.0098       2077507       2057161       2077506       2057126
    benchmark_parse_file                       -0.0110         -0.0110       2081085       2058265       2081055       2058225
    benchmark_parse_file                       -0.0097         -0.0097       2081786       2061641       2081723       2061626
    benchmark_parse_file                       -0.0110         -0.0110       2083902       2060935       2083877       2060918
    benchmark_parse_file                       -0.0100         -0.0100       2084612       2063836       2084597       2063792
    benchmark_parse_file                       -0.0099         -0.0099       2085052       2064490       2084987       2064432
    benchmark_parse_file                       -0.0118         -0.0118       2085263       2060737       2085242       2060705
    benchmark_parse_file                       -0.0080         -0.0080       2084838       2068112       2084802       2068085
    benchmark_parse_file                       -0.0120         -0.0120       2083507       2058601       2083488       2058517
    benchmark_parse_file                       -0.0107         -0.0107       2083857       2061491       2083813       2061443
    benchmark_parse_file_pvalue                 0.0002          0.0002      U Test, Repetitions: 10 vs 10
    benchmark_parse_file_mean                  -0.0104         -0.0104       2083141       2061527       2083109       2061487
    benchmark_parse_file_median                -0.0109         -0.0109       2083880       2061213       2083845       2061181
    benchmark_parse_file_stddev                +0.3608         +0.3665          2409          3278          2402          3282
    benchmark_parse_file_cv                    +0.3750         +0.3808             0             0             0             0
    OVERALL_GEOMEAN                            -0.0104         -0.0104             0             0             0             0

+++ benchmark/benchmark-utf-8.cpp
@@ -19,8 +19,8 @@ string8 repeat(string8_view s, int count) {
 void benchmark_advance_lsp_characters_in_utf_8(::benchmark::State& state,
                                                string8_view input) {
   padded_string padded_input(input);
-  int total_character_count = narrow_cast<int>(
-      count_lsp_characters_in_utf_8(&padded_input, padded_input.size()));
+  int total_character_count = narrow_cast<int>(count_lsp_characters_in_utf_8(
+      &padded_input, narrow_cast<int>(padded_input.size())));
   // Avoid count==size optimizations:
   int characters_to_count = total_character_count - 1;
